hoprd_api

Module network

Structs

AnnouncedPeerResponse 🔒

A peer that has been announced.

ConnectedPeerResponse 🔒

Immediate observation data for a connected peer.

GraphQueryRequest 🔒

Parameters for the network graph endpoint.

TicketPriceResponse 🔒

Contains the ticket price in HOPR tokens.

TicketProbabilityResponse 🔒

Contains the winning probability of a ticket.

Enums

AnnouncementOriginResponse 🔒

How a peer announcement was discovered.

Functions

announced 🔒

Lists all announced peers.

connected 🔒

Lists peers with immediate observation data from the network graph.

graph 🔒

Returns the network graph in DOT (Graphviz) format.

price 🔒

Gets the current ticket price.

probability 🔒

Gets the current minimum incoming ticket winning probability defined by the network.