hoprd_api

Module session

Structs

SessionCapabilityIter

An iterator over the variants of SessionCapability

SessionClientRequest 🔒

Request body for creating a new client session.

SessionClientResponse 🔒

Response body for creating a new client session.

SessionCloseClientQuery

SessionConfig 🔒

Enums

IpProtocol

IP transport protocol

RoutingOptions

SessionCapability

Session capabilities that can be negotiated with the target peer.

SessionTargetSpec

Session target specification.

Functions

adjust_session 🔒

close_client 🔒

Closes an existing Session listener. The listener must’ve been previously created and bound for the given IP protocol. Once a listener is closed, no more socket connections can be made to it. If the passed port number is 0, listeners on all ports of the given listening IP and protocol will be closed.

create_client 🔒

Creates a new client session returning the given session listening host and port over TCP or UDP. If no listening port is given in the request, the socket will be bound to a random free port and returned in the response. Different capabilities can be configured for the session, such as data segmentation or retransmission.

list_clients 🔒

Lists existing Session listeners for the given IP protocol.

session_config 🔒